move mocks to new pkg /pkg/aws
