+++ mock/aws/CloudWatchLogsAPI.go
